Move MaybeMasked and Unmasked to public types
